Extract app details from AppCard
